@@ -2,7 +2,7 @@ package com.coltennye.punctual.deadline.tasks;
 
 import android.content.Context;
 import android.graphics.Paint;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.AttributeSet;
 import android.widget.LinearLayout;
 import android.widget.TextView;
